Changes that Claude made

+++ DeviceTester4.cpp
@@ -19,6 +19,7 @@
 #include "Device.hpp"
 #include "Omni.hpp"
 #include "JobProperties.hpp"
+#include <cstdint>
 
 // (source setit "test XMLDevice"; ./DeviceTester4 -sproperties='XMLMasterFile="Test Me.xml"' --driver XMLOmniDevice)
 
@@ -63,7 +64,7 @@ loadDevice (PSZCRO    pszFullDeviceName,
                     "newDeviceW_JopProp_Advanced",
                     (gpointer *)&pfnNewDeviceWArgs);
 
-   if (DebugOutput::shouldOutputDeviceTester ()) DebugOutput::getErrorStream () << __FUNCTION__ << ": pfnNewDeviceWArgs = 0x" << std::hex << (int)pfnNewDeviceWArgs << std::dec << std::endl;
+   if (DebugOutput::shouldOutputDeviceTester ()) DebugOutput::getErrorStream () << __FUNCTION__ << ": pfnNewDeviceWArgs = 0x" << std::hex << (uintptr_t)pfnNewDeviceWArgs << std::dec << std::endl;
 
    if (!pfnNewDeviceWArgs)
       return 0;
